Fix Phase 2 CI failures: portable is_unique column + intra-doc links
Two regressions from a7f4392 that only surface in CI:

- `mutable_table_indexes.is_unique` was declared `INTEGER` in migration 007.
  SQLite stores that as i64-compatible variable-width; Postgres makes it
  INT4. The engine's `row.get::<i64>("is_unique")` then fails to decode on
  Postgres with `mismatched types; Rust type i64 (as SQL type INT8) is not
  compatible with SQL type INT4`. The latent bug never surfaced before
  because the prior Postgres test job never reloaded a session against a
  catalog with index rows; the new make_test_session helper triggers
  reload_mutable_tables which hits the failing decode path.
  Fix: change the column to `BIGINT`. Invisible to SQLite (stored
  identically); matches the engine's i64 read shape on Postgres.

- `cargo doc -D warnings` rejected two intra-doc links introduced in
  a7f4392: `[\`BackendImpl\`]` (no item in session.rs scope) and
  `[\`crate::JammiSession::with_backend\`]` (JammiSession is not at the
  crate root). Both rewritten to fully-qualified paths.

Verified:
- `cargo test --workspace` — 66 hermetic tests still pass
- `RUSTDOCFLAGS="-D warnings" cargo doc --workspace --no-deps` clean

+++ crates/jammi-engine/src/catalog/schema.rs
@@ -194,7 +194,10 @@ CREATE TABLE mutable_table_indexes (
     table_id        TEXT NOT NULL REFERENCES mutable_tables(id) ON DELETE CASCADE,
     index_name      TEXT NOT NULL,
     columns         TEXT NOT NULL,
-    is_unique       INTEGER NOT NULL DEFAULT 0,
+    -- BIGINT (8-byte) so the column decodes as `i64` on both SQLite (stores
+    -- INTEGER as variable-width up to 8 bytes) and Postgres (where `INTEGER`
+    -- is INT4 / i32, incompatible with the engine's i64 read shape).
+    is_unique       BIGINT NOT NULL DEFAULT 0,
     PRIMARY KEY (table_id, index_name)
 );
 "#;
